refactor: Put notifications into their own directories
